wip to mark progress before refactoring
Prep before changing cocoa bindings, etc. due to threading misunderstanding.  Observed issues: after a few hundred updates the label stops updating as fast, or at all, unless the window is moved between screens (which for some reason seems to trigger an update).   Most visible on the number of files checked label, since that updates the most.  This led to the discovery that all the controls I bound to properties in ccDirs probably need to be removed and handled differently since the bindings are implemented via KVO and based on reading, this means that every time I updated those properties in the worker threads, the observation code was executed in the same thread to update the GUI control rather than in the GUI main thread.  Same with the KVO code I added for isProcessing.  Looking into fixes, from a completely different approach to using the Receptionist pattern.