made stopping more responsive, added more error handling
Added the ability to interrupt an Apple Lossless to flac conversion, and added code to delete any partial flac files created if the conversion doesn't complete successfully.  Added some logic to the do it / stop button
and the isProcessing handler to make stopping more responsive and lock out the ability for the user to click on the do it / stop button if it's already in the process of stopping.
Removed the wait code from cancelOperations so as to not block the UI.