set up git versioning
created a new target with a script build phase which creates a .h file with defines for git tag/commit info.
Turn on preprocess Info.plist and set the preprocess header file to the created .h file.
Note that when building this won't change the Info.Plist file in the project, but rather the processed file
inside the build directory.  Since the .h file is generated each time, it can be added to .gitignore,
therefore making the whole process clean in terms of building without the build itself causing changes to tracked files