Fix sleep copy/paste code error: NoDisplaySleep ->NoIdleSleep
Oops.  Worked in testing in terms of machine not sleeping, but saw that the display
never blanked as I expected.  I had forgot to switch type of sleep to the proper one
to just keep the machine from sleeping, but still allow the display to go to screensaver,
screen power save, etc.