Delete empty tests stub from project.  wip - add playlist support, dealing with NSOutlineView
This is a checkpoint commit to save a non-working but kind of there NSOutline implementation
before yet again tearing it out and reworking it.  Bindings and tree controller didn't work as
I initially hoped, so added a data source, only to find that most of the code examples and docs
I was looking at were written for cell based, not view based NSOutlines.  It's a jumble now, but
wish to save the commit for now in order to be able to refer or roll back if needed as I reimplement.