wip checkpoint before changing playlist tree building.
New discovery - seems that playlists can conceptually contain other playlists, but it's not stored as a tree structure when we retrieve them.
instead, from experimentation it seems like the items array is only filled with media items, never other playlists.  The only way to
build the tree is to look at the parentID of a playlist.  A playlist itself has no concept of it's children, only the child knows it's parent.
Furthermore, in iTunes you actually create either playlists or playlist folders.  A playlist folder can only contain other playlists, and
appears to be represented as a playlist with a nil items array.
Based on this, I need to redo the tree structure code, since I should never have to look at the items array, and will probably just build a
dictionary based tree.