Changed version fields for Sparkle usage
Sparkle uses the CFBundleVersion (the BUILD choice, in Xcode's property pages) to compare with the appcast's version field.
It expects it to be an increasing value, and handles version component separators like '.'  However it wouldn't have any idea
what to do with a git commit ref, which is what I previously used for the BUILD number.  Since the GIT_VERSION string I create
should conform to Sparkle's expectations I switched the CFBundleVersion to that, then took out the CFBundleShortVersionString
to keep the about box of the app from showing the same data twice, once in parenthesis.