Add options for "remap album artist / artist" and "strip tags for playlists".  Issues: fix #2, fix #3.
Adds the ability to remap album artist to artist, and append artist to the song title.
This option can help with compilation albums getting "fragmented" in the play by album view on the Model S.
Adds the ability to strip title, album, artist, and genre from playlist items (though genre will still be set to the
playlist name if that option is checked).  This might help avoid duplicate entries in the other play modes,
though you'll lose the display of the tags in the play by genre (if that option is checked) or play by folder views.
Since the Model S doesn't currently display any metadata from the song in the play by folder view, that's not
much of a loss.
While adding this functionality, a couple of bugs were discovered and fixed relating to alac->flac conversion genre
tag setting, and a playlist cleanup bug leading to a race condition where the app would try to delete the destination
playlist folders when it should not.