Security fix for Sparkle MITM vulnerability
Changed Sparkle update URLs to all be https.  While the app was already
signed and the download URLs were already https, the appcast.xml itself
was http. A man in the middle attack was discovered against this scenario.

This change fixes the MITM issue, even if the Sparkle component itself
isn't updated.  Will update the Sparkle component as well next, in order to
also close a code injection vulnerability discussed in the same security
note.  With the MITM vulnerability closed via https, the latter issue could
only be exploited if an attacker had the ability to compromise the appcast
xml file itself on the server.

Also, with El Capitan, the OS enforces TLS by default, so automatic
updates weren't working, um, automatically.  On the plus side, that made
the Sparkle MITM attack impossible, but with these changes should work again.

Recompiled with Xcode 7.2.1 - some project settings and layout files got
updated in the process.  Hopefully everything still works on prior versions.