added ATS exceptions to enable Sparkle updates on OS X 10.11+
Added exceptions to allow updating to continue even though it might end
up not being served through HTTPS.  This is suboptimal, but the app cast
and files are currently served via github, and they do not reliably support
HTTPS for github pages type connections.  In particular even when using
https://whateverat.github.io/app/appcast.xml I see the server redirect me
to a variant of my custom domain URL using http.
These exceptions will be removed if/when either github starts supporting
https in a way I can use, or I find a different hosting provider I can leverage
without paying much if anything for (given this is a free project).