Updates project settings & includes frameworks
Cleaned project settings of fixed paths, removed some cruft, and added a
frameworks directory containing the version of Sparkle and taglib the
project is built with.  The remaining non-included, non-standard
frameworks and libraries are the iTunes framework from Apple and the
FLAC static library, both of which are installed on my system in system
locations - though I suppose I should really include them in the project
as well, or note the versions.