Recompile framework with deployment target set to 10.10 so that the app will still work on 10.10.x
A user noted that the app crashes when on OS X 10.10.5.  There is a dyld loading error, code signature fail, despite the signing looking good on newer versions.  After digging, seems to be that the code signing process looks at the framework's deployment target rather than the app target when determining which signing options to use. See:
  https://stackoverflow.com/questions/40920801/failure-when-trying-to-run-application-release-version-on-another-machine-with-t/40980658#40980658
  https://stackoverflow.com/questions/17143373/determine-minimum-osx-version-a-binary-was-compiled-for/17148833#17148833
  https://stackoverflow.com/questions/41865537/how-does-apples-codesign-utility-decide-which-sha-algorithms-to-sign-a-shared

Quick fix was to recompile taglib with cmake's -DCMAKE_OSX_DEPLOYMENT_TARGET=10.10 set